Updated Org-mode configuration for Emacs
Added hook for flyspell-mode, disabled emphasis hiding, removed fancy
font configuration because it's better when it's simpler.
I also put a lot of code into a function which is called when a frame
is created, because a window error is generated when lauching the
Emacs daemon.
Removed validation button at the end of Org-mode's HTML exports.